+++ node/commons/compatibility-layer.ts
@@ -28,7 +28,7 @@ export const convertBiggyProduct = (
     convertSKU(product, indexingType, tradePolicy)
   )
 
-  const selectedProperties = [
+  const selectedProperties = product.split && [
     {
       key: product.split.labelKey,
       value: product.split.labelValue,