+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Update `vtex.search-graphql` version.
+
 ## [0.8.4] - 2020-07-09
 ### Fixed
 - Translate variations fields inside the SKU items with id as well.