+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Deal with `null` exception in `product.split`.
+
 ## [1.8.0] - 2020-07-09
 ### Added
 - `selectedProperties` to the product object.